Ver 8.2  8.1 notwork  change long to BigInteger in Domain.java

+++ sample-spring-boot-kitchensink/src/main/java/com/example/bot/spring/Domain.java
@@ -4,22 +4,24 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
+
 @Document(collection = "domain")
 public class Domain {
 
     @Id
-    private long id;
+    private BigInteger id;
 
     @Indexed(unique = true)
     private String domain;
 
     private boolean displayAds;
 
-    public long getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
